Migrate legacy tests to Bun
